backend/requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2024.2

# ASGI & Channels
asgiref==3.8.1
channels==4.2.0

# Security & CORS
django-cors-headers==4.6.0
PyJWT==2.10.1

# Data Processing
pandas==2.2.2
pyarrow==16.1.0
fastparquet==2024.2.0
numpy<2

# Machine Learning
torch==2.2.2
scikit-learn==1.6.1
joblib==1.3.2
pytorch_tabnet
Pillow
psutil

# Network Analysis
pyshark==0.6
requests==2.31.0